Added multithreaded image file export (cutting game time by roughly 25%). Multithreading of logger still not working

+++ bit-mould/Definitions/Image/Color.h
@@ -1,6 +1,7 @@
 #pragma once
+
 struct Color {
-	float r, g, b;
+	long double r, g, b;
 	static const int size = 3;
 
 	Color();
